+++ include/CpuMeasure.h
@@ -12,7 +12,7 @@
  */
 struct CpuMeasure {
     /// @brief Time of the measure in seconds since start.
-    unsigned long long time_seconds;
+    double time_seconds;
 
     /// @brief CPU percentage of this measure.
     double percentage;
